requirements.txt
# requirements.txt
streamlit>=1.31
pandas>=2.0
numpy>=1.24
requests>=2.31
python-dotenv>=1.0
beautifulsoup4>=4.12
lxml>=5.0
